# [WIP] Simplify and speed-up numpy smallest-k