In the current numpy-based implementation of top-k, we use `np.argpartition` to get the smallest `k` scores. We then do a resort because we assumed that the returned partition might not be in sorted order (discussed [here](https://github.com/awslabs/sockeye/issues/270)). It turns out that `np.argpartition` can return an already sorted partition, but it's a bit hidden in the docs, found out about it on [stack overflow](https://stackoverflow.com/a/42185645).

In microbenchmarks, it's nearly 2x faster that the current approach in master:

```
In [2]: import numpy as np

In [3]: long = np.random.rand(5, 32000)

In [4]: long
Out[4]:
array([[ 0.19846017,  0.95168003,  0.6515433 , ...,  0.23239723,
         0.05825158,  0.19316375],
       [ 0.56247234,  0.12604064,  0.63571076, ...,  0.2767887 ,
         0.7993372 ,  0.64724466],
       [ 0.19306748,  0.06372986,  0.90068909, ...,  0.70291493,
         0.65553369,  0.01743039],
       [ 0.00679545,  0.83289007,  0.09616722, ...,  0.92498974,
         0.19441304,  0.68613256],
       [ 0.98585787,  0.56107536,  0.63116259, ...,  0.10633627,
         0.43929976,  0.51507055]])

In [5]: long = long.flatten()

In [6]: long.shape
Out[6]: (160000,)

In [10]: %%timeit
    ...: args = np.argpartition(long, 5)[:5];
    ...: args = args[np.argsort(long[args])]
    ...:
1.89 ms ± 197 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [11]: %%timeit
    ...: args = np.argpartition(long, range(5))[:5]
    ...:
977 µs ± 88.1 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

I also translated a large (5431 sent) test set on GPU on a p2.xlarge and compared the total times:

* Mxnet-based top-k: `Total time: 954.4973, sec/sent: 0.1757, sent/sec: 5.6899`
* Numpy-based top-k: `Total time: 1031.7326, sec/sent: 0.1900, sent/sec: 5.2640`
* Simplified numpy top-k (this PR): `Total time: 946.1822, sec/sent: 0.1742, sent/sec: 5.7399`

There were no differences in the translation output.

Given that the numbers look better than the mxnet-based one, I would recommend switching over to the simplified numpy version pending further speed tests. I'm currently running the same test on CPU, but it will take a while. I'll update the PR with the results. I think @mjpost has a nice setup for speed testing, would be nice if you could run the code in this PR :)

#### Pull Request Checklist ##
- [ ] Changes are complete (if posting work-in-progress code, prefix your pull request title with '[WIP]'
until you can check this box.
- [ ] Unit tests pass (`pytest`)
- [ ] Were system tests modified? If so did you run these at least 5 times to account for the variation across runs?
- [ ] System tests pass (`pytest test/system`)
- [ ] Passed code style checking (`./style-check.sh`)
- [ ] You have considered writing a test
- [ ] Updated major/minor version in `sockeye/__init__.py`. Major version bump if this is a backwards incompatible change.
- [ ] Updated CHANGELOG.md


By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.