@fhieber 2018-03-26
I love these kinds of small but relevant changes, great finding, thanks!
Let's wait for the CPU times and have @mjpost chime in on whether we should revert the smallest_k_mx change (using mxnet topk) we just introduced.