@Deseaus 2018-03-27
A small update with initial numbers on CPU:

* Numpy master: `Total time: 5022.9249, sec/sent: 0.9249, sent/sec: 1.0812`
* Numpy simplified: `Total time: 4993.4917, sec/sent: 0.9194, sent/sec: 1.0876`

Slight improvement, this was only run on a p2.xlarge though, no exactly the most powerful CPU machine :)