@fhieber 2018-03-27
One thing we could do is to decouple the revert of the cpu/gpu branching and only change the utils.smallest_k() function. We could then merge this PR as is and revisit the context branching again later.