@Deseaus 2018-03-27
I pushed a new version where I'm batching the top_k computation in a single call. While this doesn't make much of a difference in single-sentence translation, it does bring a nice speed-up in batch translation. I've consolidated the various scores below (*simplified numpy top-k* was the first iteration of this PR adding `range()` to `argpartition`; *batched numpy top-k* is the current version).

The code is a mess and would need to be cleaned up, but just so you get an idea of the approach. I haven't found a cleaner way of offsetting the indexes in `best_hyp_indices`. Still needs further testing for sure to find the overall speed differences. At batch size 10, there doesn't seems to be much of a difference between *simplified numpy top-k* and *master numpy-based top-k*, for example.

## GPU results:

```
Master mxnet-based top-k     [GPU] [Batch 1]: `Total time: 954.4973, sec/sent: 0.1757, sent/sec: 5.6899`
Master numpy-based top-k     [GPU] [Batch 1]: `Total time: 1031.7326, sec/sent: 0.1900, sent/sec: 5.2640`
Simplified numpy top-k       [GPU] [Batch 1]: `Total time: 946.1822, sec/sent: 0.1742, sent/sec: 5.7399`
Batched numpy top-k          [GPU] [Batch 1]: `Total time: 932.8663, sec/sent: 0.1718, sent/sec: 5.8218`
```

```
Master mxnet-based top-k       [GPU] [Batch 10]:  ???
Master numpy-based top-k       [GPU] [Batch 10]: `Total time: 452.3622, sec/sent: 0.0833, sent/sec: 12.0059`
Simplified numpy top-k         [GPU] [Batch 10]: `Total time: 452.3817, sec/sent: 0.0833, sent/sec: 12.0053`
Batched numpy top-k            [GPU] [Batch 10]: `Total time: 273.6624, sec/sent: 0.0504, sent/sec: 19.8456`
```

## CPU results:

```
Master numpy-based top-k     [CPU] [Batch 1]: `Total time: 5022.9249, sec/sent: 0.9249, sent/sec: 1.0812`
Simplified numpy top-k       [CPU] [Batch 1]: `Total time: 4993.4917, sec/sent: 0.9194, sent/sec: 1.0876`
Batched numpy top-k          [CPU] [Batch 1]: ???
```