@fhieber 2018-03-28
I too the freedom to take your code and merge it with what I had for batched mxnet topk. How about adding a function in utils.py that looks like this:
```python
def topk(scores: mx.nd.NDArray, k: int, batch_size: int, t: int, use_mxnet_topk: bool):
    # scores: (batch_size * beam_size, target_vocab_size)
    if t == 1:  # beam size is effectively 1
        folded_scores = scores[::k, :]
    else:
        # (batch_size, beam_size * vocab_size)
        folded_scores = scores.reshape((batch_size, k * scores.shape[-1]))

    if use_mxnet_topk:
        values, indices = mx.nd.topk(folded_scores, axis=1, k=k, ret_typ='both', is_ascend=True)
        best_hyp_indices, best_word_indices = np.unravel_index(indices.astype(np.int32).asnumpy().ravel(), scores.shape)
        values = values.reshape((-1,))
    else:
        folded_scores = folded_scores.asnumpy()
        # Get the scores
        # Indexes into folded_scores: (batch_size, beam_size)
        flat_idxs = np.argpartition(folded_scores, range(k))[:, :k]
        # Score values: (batch_size, beam_size)
        values = folded_scores[np.arange(folded_scores.shape[0])[:, None], flat_idxs].ravel()
        best_hyp_indices, best_word_indices = np.unravel_index(flat_idxs.ravel(), scores.shape)
```

This entirely gets rid of the previous for loop and would be called in `_beam_search()` like this:
```python
best_hyp_indices[:], best_word_indices[:], scores_accumulated[:, 0] = utils.topk(scores, self.beam_size, self.batch_size, t=t, use_mxnet_topk=False)  # TODO: determine when to switch
```

We could use something like this to benchmark all variations (CPU/GPU, batch/non-batch) etc.