@Deseaus 2018-04-11
I ran some speed tests on a p2.xlarge on GPU with a variety of setups, here are the graphs:

* The green and orange lines correspond to baselines in master, with the commit ID in brackets.
* Red and blue are the batched topK implementations that @fhieber referenced above.
* The purple line is a two-step topK explained below.

## Split by batch size

![speedbybatch](https://user-images.githubusercontent.com/6504048/38606829-71bfad28-3d77-11e8-83b4-e4337d89690b.png)


## Split by beam size

![speedbybeam](https://user-images.githubusercontent.com/6504048/38606838-753cc706-3d77-11e8-95c7-9f222dad2889.png)

## Discussion

* It seems that on GPU the batched topK implementations don't negatively affect the batch=1 use case, as was expected.
* Batched topK brings nice speed-ups for most batch-beam size combinations, with some coming in roughly the same as the baseline.
* Mxnet and numpy batched topK versions are not terribly different, with the mxnet version performing better for `beam_size>5` and the numpy version for `beam_size<5`.
* The batched topK versions show a nearly linear speed-up as batch sizes increases in the greedy-case (batch=1). 

This last point led me to think that the bottleneck in the topK was the larger vocabulary dimension, which wasn't scaling well. For larger beam sizes we'd be performing topK row-wise on a matrix of size `(batch_size, beam_size * vocab_size)`, where adding more rows (i.e. sentences) scaled well as seen in the graphs for beam_size=1. So I tried the purple line, which is a two-step topK: first run topK row-wise on the matrix `(batch_size * beam_size, vocab_size)`, so we'd end up with the top k entries for each *hypothesis*. Then, run another topK on the reshaped result, which would only be `(batch_size, k*k)`. Unfortunately, this didn't show much improvement.

All in all, it seems the batched topK operators work well on GPU decoding, leaning towards the mxnet one. We should check there's no degradation on CPU for non-batched decoding.