@fhieber 2018-04-11
Fantastic, thanks a lot for this extensive benchmarking!!
Based on your first picture, I would go for using mxnet-based batched topk on GPUs as it outperforms at beam size 5 the numpy version (beam size 5 being a common default setting). 
Can you run the CPU benchmarks as well? I think we should then go ahead and implement this asap.