@Deseaus 2018-04-12
I'm running some CPU benchmarks now on a c5.xlarge. Only with batch sizes 1 and 10 now, since batching doesn't make much sense on a CPU. But wanted to try batch 10 and see what it looks like :)